==> notebooks/04_nam_clean_data_us_embass.ipynb <==
from __future__ import print_function
import os
import sys
PROJ_ROOT = os.path.join(os.pardir)
print(os.path.abspath(PROJ_ROOT))
src_dir = os.path.join(PROJ_ROOT, "src")
sys.path.append(src_dir)
%load_ext autoreload

import pandas as pd
import numpy as np
%aimport features.calculate_AQI
from features import calculate_AQI

# We need to know the number of continuous values
# So when we use chunker, we can refer to the starting point to know how many to skip if timeline is not continous after skipping
# Basic idea: Loop through the dataframe, if it is continuous increase the continous counter and pos counter
# If encounter uncontinous timeline (Encounter a nan value), find the lastpoint with a loop,
# Then change the start position continous length column values to the continous counter.
# Reset the continous counter and start counter
# Result: at the start of each continous timeline, there will be a number which indicate how many step you need to skip if
# you need to get to the next timeline

def time_continous_marker(df):
    df_copy = df.copy()
    pos = 0
    start_pos = 0
    continous = 0
    size = 24
    length = df_copy.shape[0]
    result_column = df_copy.columns.get_loc("Continous length")
    continous_column = df_copy.columns.get_loc("AQI_h")
    while pos < length:
        try:
            if df_copy.iloc[pos : pos + size, continous_column].isnull().sum() == 0:
                while df_copy.iloc[pos : pos + size, continous_column].isnull().sum() == 0:
                    pos += size
                    continous += size
            else:
                while np.isnan(df_copy.iloc[pos, continous_column]) == False:
                    continous += 1
                    pos += 1
                
                df_copy.iloc[start_pos, result_column] = continous
                continous = 0
                while np.isnan(df_copy.iloc[pos, 1]) == True:
                    pos += 1
                start_pos = pos
#             print(pos)
        except IndexError:
            df_copy.iloc[start_pos, result_column] = continous
            continous = 0
            print("Current position is: {}".format(pos))
            pos = length
    return df_copy

# TP HCM
import pandas as pd
import numpy as np
import glob
import xarray as xr

idx = pd.IndexSlice

_HCM_us_embass_raw_data_path = os.path.join(PROJ_ROOT,
                             "data",
                             "external",
                             "us_embass",
                             "raw")

_HCM_us_embass_data_files = glob.glob(_HCM_us_embass_raw_data_path + '/HoChiMinhCity_PM2.5_*YTD.csv')
USEm_HCM_df = pd.DataFrame()
for file in _HCM_us_embass_data_files:
    USEm_HCM_df = USEm_HCM_df.append(pd.read_csv(file, parse_dates=True))

# Remove site year 2015 because inconsistency data
# daisuquanmy_df = daisuquanmy_df[(daisuquanmy_df['Year'] != 2015)]
USEm_HCM_df.drop(columns=['Site', 'Parameter', 'AQI', 'AQI Category','NowCast Conc.', 'Conc. Unit', 'Duration'], inplace=True)
USEm_HCM_df['Date (LT)'] = pd.to_datetime(USEm_HCM_df['Date (LT)'])
USEm_HCM_df = USEm_HCM_df.rename(columns={'Date (LT)': 'time'})
# Remove duplicate, fill in missing index and set ['Date (LT)'] as index and sort
USEm_HCM_df.drop_duplicates(subset ="time", 
                     keep = "first", inplace = True)
USEm_HCM_df = USEm_HCM_df.sort_values(by=['time'])
USEm_HCM_df = USEm_HCM_df.set_index('time').asfreq('H').sort_index()
USEm_HCM_df.reset_index(drop=False, inplace=True)
USEm_HCM_df['site_id'] = 49
USEm_HCM_df = USEm_HCM_df.set_index(['site_id', 'time'])
# For all Raw Conc <= 0 QC Name must be Missing. Raw Conc change to -999
USEm_HCM_df.loc[USEm_HCM_df['Raw Conc.'] <= 0, 'Raw Conc.'] = -999
USEm_HCM_df.loc[USEm_HCM_df['Raw Conc.'] >= 600, 'Raw Conc.'] = -999
USEm_HCM_df.loc[USEm_HCM_df['Raw Conc.'] <= 0, 'QC Name'] = 'Missing'

USEm_HCM_df = USEm_HCM_df.replace(-999, np.nan)
# Change Raw Conc to PM25
USEm_HCM_df.rename(columns={'Raw Conc.': 'PM25'}, inplace=True)
# Fill missing data
USEm_HCM_df.fillna(method='ffill', limit=2, inplace=True)
USEm_HCM_df.fillna(method='bfill', limit=2, inplace=True)

# For year 2015 only, take data after 2015-12-09 14:00:00
USEm_HCM_df = USEm_HCM_df.loc[USEm_HCM_df.index.get_level_values(1) >= pd.to_datetime("2016-02-05 02:00 PM")]
USEm_HCM_df = USEm_HCM_df[['PM25']]

# USEm_HCM_df['PM25'] = USEm_HCM_df['PM25'].interpolate()

USEm_HCM_df

AQI = calculate_AQI.calculate_AQI_h(USEm_HCM_df)
AQI_new = AQI.copy()
AQI_new.loc[AQI['AQI_h'] <= 0, 'AQI_h'] = np.NAN
AQI_new['Continous length'] = 0
max_time = AQI_new.index.get_level_values(1).max() + pd.Timedelta(hours=1)
AQI_new = AQI_new.append(pd.Series(name=(49, max_time), dtype=float))
AQI_new = time_continous_marker(AQI_new)
AQI_new = AQI_new.dropna(subset=['AQI_h'])
AQI_new

_HCM_us_embass_processed_data_path = os.path.join(PROJ_ROOT,
                             "data",
                             "external",
                             "us_embass",
                             "processed")
AQI_new.to_csv(_HCM_us_embass_processed_data_path + "/hcm_processed.csv")

# Ha Noi
import pandas as pd
import numpy as np
import glob
import xarray as xr

idx = pd.IndexSlice

_HCM_us_embass_raw_data_path = os.path.join(PROJ_ROOT,
                             "data",
                             "external",
                             "us_embass",
                             "raw")

_Hanoi_us_embass_data_files = glob.glob(_HCM_us_embass_raw_data_path + '/Hanoi_PM2.5_*YTD.csv')
USEm_Hanoi_df = pd.DataFrame()
for file in _Hanoi_us_embass_data_files:
    USEm_Hanoi_df = USEm_Hanoi_df.append(pd.read_csv(file, parse_dates=True))

# Remove site year 2015 because inconsistency data
# daisuquanmy_df = daisuquanmy_df[(daisuquanmy_df['Year'] != 2015)]
USEm_Hanoi_df.drop(columns=['Site', 'Parameter', 'AQI', 'AQI Category','NowCast Conc.', 'Conc. Unit', 'Duration'], inplace=True)
USEm_Hanoi_df['Date (LT)'] = pd.to_datetime(USEm_Hanoi_df['Date (LT)'])
USEm_Hanoi_df = USEm_Hanoi_df.rename(columns={'Date (LT)': 'time'})
# Remove duplicate, fill in missing index and set ['Date (LT)'] as index and sort
USEm_Hanoi_df.drop_duplicates(subset ="time", 
                     keep = "first", inplace = True)
USEm_Hanoi_df = USEm_Hanoi_df.sort_values(by=['time'])
USEm_Hanoi_df = USEm_Hanoi_df.set_index('time').asfreq('H').sort_index()
USEm_Hanoi_df.reset_index(drop=False, inplace=True)
USEm_Hanoi_df['site_id'] = 49
USEm_Hanoi_df = USEm_Hanoi_df.set_index(['site_id', 'time'])
# For all Raw Conc <= 0 QC Name must be Missing. Raw Conc change to -999
USEm_Hanoi_df.loc[USEm_Hanoi_df['Raw Conc.'] <= 0, 'Raw Conc.'] = -999
USEm_Hanoi_df.loc[USEm_Hanoi_df['Raw Conc.'] >= 600, 'Raw Conc.'] = -999
USEm_Hanoi_df.loc[USEm_Hanoi_df['Raw Conc.'] <= 0, 'QC Name'] = 'Missing'

USEm_Hanoi_df = USEm_Hanoi_df.replace(-999, np.nan)
# Change Raw Conc to PM25
USEm_Hanoi_df.rename(columns={'Raw Conc.': 'PM25'}, inplace=True)
# Fill missing data
USEm_Hanoi_df.fillna(method='ffill', limit=2, inplace=True)
USEm_Hanoi_df.fillna(method='bfill', limit=2, inplace=True)

# For year 2015 only, take data after 2015-12-09 14:00:00
USEm_Hanoi_df = USEm_Hanoi_df.loc[USEm_Hanoi_df.index.get_level_values(1) >= pd.to_datetime("2015-12-09 14:00:00")]
USEm_Hanoi_df = USEm_Hanoi_df[['PM25']]

# USEm_HCM_df['PM25'] = USEm_HCM_df['PM25'].interpolate()

USEm_Hanoi_df

AQI = calculate_AQI.calculate_AQI_h(USEm_Hanoi_df)
AQI_new = AQI.copy()
AQI_new.loc[AQI['AQI_h'] <= 0, 'AQI_h'] = np.NAN
AQI_new['Continous length'] = 0
max_time = AQI_new.index.get_level_values(1).max() + pd.Timedelta(hours=1)
AQI_new = AQI_new.append(pd.Series(name=(49, max_time), dtype=float))
AQI_new = time_continous_marker(AQI_new)
AQI_new = AQI_new.dropna(subset=['AQI_h'])
AQI_new

_Hanoi_us_embass_processed_data_path = os.path.join(PROJ_ROOT,
                             "data",
                             "external",
                             "us_embass",
                             "processed")
AQI_new.to_csv(_Hanoi_us_embass_processed_data_path + "/hanoi_processed.csv")

USEm_Hanoi_df.to_csv(_Hanoi_us_embass_processed_data_path + "/hanoi_conc.csv")

# Now chunker need to be able to output chunks of train data and target in correct order
# Basic ideas: loop through each chunk of data (size is size with ,
# if the timeline is continous, add data to chunk_data [pos: pos+size]
# add data to target [pos+size+target_hour]
# if the timeline is not, skip pos according to size
# Repeat below till timeline is continous
# if after skipping, the timeline is still not continous, skip pos according to continous length colunmns
from IPython.core.debugger import set_trace
def chunker(seq, size, target_hour, debug = True):
    """
    Input: 
        + Dataframe with PM25, AQI_h, AQI_h_I, Continous length and some other columns
            - If data doesn't have Continous length columns, add 1 to data
            - Take only PM25, date columns as train
            - Take only AQI_h, AQI_h_I as target
        + size: length of each chunk
        + target_hour: labels for the hour which will be predicted
    Ouput: 2 lists: chunk_data, target
        + chunk_data: original data chunked to a specific timestep, have shape [..., timeframe, features]
        + target: Label for each chunked train/test data, have shape [..., 1]
    """
    if not 'Continous length' in seq.columns:
        seq['Continous length'] = 0

    timerange = pd.Timedelta(hours=size + target_hour)
    chunk_data = []
    target = []
    multiclass_target = []
    pos = 0
    length = seq.shape[0]
    while pos < (length - size):
        try:
            if seq.iloc[pos + size + target_hour].name[1] - seq.iloc[pos].name[1] == timerange:
                while seq.iloc[pos + size + target_hour].name[1] - seq.iloc[pos].name[1] == timerange:
                    chunk_data.append(seq.iloc[pos:pos + size].loc['PM25','Hour','Day of Week'])
                    target.append(seq.iloc[pos + size + target_hour].loc['AQI_h'])
                    multiclass_target.append(seq.iloc[pos + size + target_hour].loc['AQI_h_I'])
                    pos += 1
                    if pos + size + target_hour >= length - size:
                        print("Returned")
                        return np.array(chunk_data), np.array(target), np.array(multiclass_target)
            else:
                tmp_pos = pos + size + target_hour
                while seq.iloc[tmp_pos + size + target_hour].name[1] - seq.iloc[tmp_pos].name[1] != timerange:
                    tmp_pos += int(seq.iloc[tmp_pos, 6])
                print("Jump from {} to {}".format(pos, tmp_pos))
                pos = tmp_pos
        except IndexError:
            print("Current position is: {}".format(pos))
            print("Current tmp position is: {}".format(tmp_pos))
            pos = length - size
    print("Returned")
    return np.array(chunk_data), np.array(target), np.array(multiclass_target)
chunk_data, target, multiclass_target = chunker(AQI_copy, 12, 0)
